Adding AbstractRadarLoginManager for common logic of logic managers
