+++ radar-commons-android/src/main/java/org/radarbase/android/auth/portal/ManagementPortalLoginManager.kt
@@ -2,7 +2,6 @@ package org.radarbase.android.auth.portal
 
 import android.app.Activity
 import androidx.lifecycle.Observer
-import org.json.JSONException
 import org.radarbase.android.RadarApplication.Companion.radarConfig
 import org.radarbase.android.RadarConfiguration.Companion.MANAGEMENT_PORTAL_URL_KEY
 import org.radarbase.android.RadarConfiguration.Companion.OAUTH2_CLIENT_ID
@@ -10,23 +9,24 @@ import org.radarbase.android.RadarConfiguration.Companion.OAUTH2_CLIENT_SECRET
 import org.radarbase.android.RadarConfiguration.Companion.UNSAFE_KAFKA_CONNECTION
 import org.radarbase.android.auth.AppAuthState
 import org.radarbase.android.auth.AuthService
-import org.radarbase.android.auth.LoginManager
-import org.radarbase.android.auth.SourceMetadata
+import org.radarbase.android.auth.commons.AbstractRadarLoginManager
+import org.radarbase.android.auth.commons.AbstractRadarPortalClient
 import org.radarbase.android.auth.portal.ManagementPortalClient.Companion.MP_REFRESH_TOKEN_PROPERTY
 import org.radarbase.android.util.ServerConfigUtil.toServerConfig
 import org.radarbase.android.config.SingleRadarConfiguration
 import org.radarbase.config.ServerConfig
-import org.radarbase.producer.AuthenticationException
 import org.radarbase.producer.rest.RestClient
 import org.slf4j.LoggerFactory
-import java.io.IOException
 import java.net.MalformedURLException
 import java.util.concurrent.locks.ReentrantLock
+import kotlin.contracts.ExperimentalContracts
+import kotlin.contracts.contract
 
-class ManagementPortalLoginManager(private val listener: AuthService, state: AppAuthState) : LoginManager {
-    private val sources: MutableMap<String, SourceMetadata> = mutableMapOf()
+@Suppress("unused")
+class ManagementPortalLoginManager(private val listener: AuthService, state: AppAuthState) :
+    AbstractRadarLoginManager(listener) {
 
-    private var client: ManagementPortalClient? = null
+    override var client: AbstractRadarPortalClient? = null
     private var clientConfig: ManagementPortalConfig? = null
     private var restClient: RestClient? = null
     private val refreshLock: ReentrantLock
@@ -47,27 +47,31 @@ class ManagementPortalLoginManager(private val listener: AuthService, state: App
 
     fun setTokenFromUrl(authState: AppAuthState, refreshTokenUrl: String) {
         client?.let { client ->
-            if (refreshLock.tryLock()) {
-                try {
-                    // create parser
-                    val parser = MetaTokenParser(authState)
+            if (ensureMpClient(client)) {
+                if (refreshLock.tryLock()) {
+                    try {
+                        // create parser
+                        val parser = MetaTokenParser(authState)
 
-                    // retrieve token and update authState
-                    client.getRefreshToken(refreshTokenUrl, parser).let { authState ->
-                        // update radarConfig
-                        config.updateWithAuthState(listener, authState)
+                        // retrieve token and update authState
+                        client.getRefreshToken(refreshTokenUrl, parser).let { authState ->
+                            // update radarConfig
+                            config.updateWithAuthState(listener, authState)
                             // refresh client
-                        ensureClientConnectivity(config.latestConfig)
-                        logger.info("Retrieved refreshToken from url")
-                        // refresh token
-                        refresh(authState)
+                            ensureClientConnectivity(config.latestConfig)
+                            logger.info("Retrieved refreshToken from url")
+                            // refresh token
+                            refresh(authState)
+                        }
+                    } catch (ex: Exception) {
+                        logger.error("Failed to get meta token", ex)
+                        listener.loginFailed(this, ex)
+                    } finally {
+                        refreshLock.unlock()
                     }
-                } catch (ex: Exception) {
-                    logger.error("Failed to get meta token", ex)
-                    listener.loginFailed(this, ex)
-                } finally {
-                    refreshLock.unlock()
                 }
+            } else {
+                logger.error("Invalid client type detected: expected ManagementPortalClient for the meta token URL. Terminating operation")
             }
         }
     }
@@ -77,22 +81,26 @@ class ManagementPortalLoginManager(private val listener: AuthService, state: App
             return false
         }
         client?.let { client ->
-            if (refreshLock.tryLock()) {
-                try {
-                    val parser = SubjectTokenParser(client, authState)
+            if (ensureMpClient(client)) {
+                if (refreshLock.tryLock()) {
+                    try {
+                        val parser = SubjectTokenParser(client, authState)
 
-                    client.refreshToken(authState, parser).let { authState ->
-                        logger.info("Refreshed JWT")
+                        client.refreshToken(authState, parser).let { authState ->
+                            logger.info("Refreshed JWT")
 
-                        updateSources(authState)
-                        listener.loginSucceeded(this, authState)
+                            updateSources(authState)
+                            listener.loginSucceeded(this, authState)
+                        }
+                    } catch (ex: Exception) {
+                        logger.error("Failed to get access token", ex)
+                        listener.loginFailed(this, ex)
+                    } finally {
+                        refreshLock.unlock()
                     }
-                } catch (ex: Exception) {
-                    logger.error("Failed to get access token", ex)
-                    listener.loginFailed(this, ex)
-                } finally {
-                    refreshLock.unlock()
                 }
+            } else {
+                logger.error("Invalid client type detected: expected ManagementPortalClient for the meta token URL. Terminating operation.")
             }
         }
         return true
@@ -125,133 +133,6 @@ class ManagementPortalLoginManager(private val listener: AuthService, state: App
 
     override val sourceTypes: List<String> = sourceTypeList
 
-    override fun updateSource(appAuth: AppAuthState, source: SourceMetadata, success: (AppAuthState, SourceMetadata) -> Unit, failure: (Exception?) -> Unit): Boolean {
-        logger.debug("Handling source update")
-
-        val client = client
-
-        if (client != null) {
-            try {
-                val updatedSource = client.updateSource(appAuth, source)
-                success(addSource(appAuth, updatedSource), updatedSource)
-            } catch (ex: UnsupportedOperationException) {
-                logger.warn("ManagementPortal does not support updating the app source.")
-                success(addSource(appAuth, source), source)
-            } catch (ex: ManagementPortalClient.Companion.SourceNotFoundException) {
-                logger.warn("Source no longer exists - removing from auth state")
-                val updatedAppAuth = removeSource(appAuth, source)
-                registerSource(updatedAppAuth, source, success, failure)
-            } catch (ex: ManagementPortalClient.Companion.UserNotFoundException) {
-                logger.warn("User no longer exists - invalidating auth state")
-                listener.invalidate(appAuth.token, false)
-                failure(ex)
-            } catch (ex: java.lang.IllegalArgumentException) {
-                logger.error("Source {} is not valid", source)
-                failure(ex)
-            } catch (ex: AuthenticationException) {
-                listener.invalidate(appAuth.token, false)
-                logger.error("Authentication error; failed to update source {} of type {}",
-                    source.sourceName, source.type, ex)
-                failure(ex)
-            } catch (ex: IOException) {
-                logger.error("Failed to update source {} with {}",
-                    source.sourceName, source.type, ex)
-                failure(ex)
-            } catch (ex: JSONException) {
-                logger.error("Failed to update source {} with {}",
-                    source.sourceName, source.type, ex)
-                failure(ex)
-            }
-        } else {
-            failure(IllegalStateException("Cannot update source without a client"))
-        }
-
-        return true
-    }
-
-    override fun registerSource(authState: AppAuthState, source: SourceMetadata,
-                                success: (AppAuthState, SourceMetadata) -> Unit,
-                                failure: (Exception?) -> Unit): Boolean {
-        logger.debug("Handling source registration")
-
-        val existingSource = sources[source.sourceId]
-        if (existingSource != null) {
-            success(authState, existingSource)
-            return true
-        }
-
-        client?.let { client ->
-            try {
-                val updatedSource = if (source.sourceId == null) {
-                    // temporary measure to reuse existing source IDs if they exist
-                    source.type?.id
-                            ?.let { sourceType -> sources.values.find { it.type?.id == sourceType } }
-                            ?.let {
-                                source.sourceId = it.sourceId
-                                source.sourceName = it.sourceName
-                                client.updateSource(authState, source)
-                            }
-                            ?: client.registerSource(authState, source)
-                } else {
-                    client.updateSource(authState, source)
-                }
-                success(addSource(authState, updatedSource), updatedSource)
-            } catch (ex: UnsupportedOperationException) {
-                logger.warn("ManagementPortal does not support updating the app source.")
-                success(addSource(authState, source), source)
-            } catch (ex: ManagementPortalClient.Companion.SourceNotFoundException) {
-                logger.warn("Source no longer exists - removing from auth state")
-                val updatedAuthState = removeSource(authState, source)
-                registerSource(updatedAuthState, source, success, failure)
-            } catch (ex: ManagementPortalClient.Companion.UserNotFoundException) {
-                logger.warn("User no longer exists - invalidating auth state")
-                listener.invalidate(authState.token, false)
-                failure(ex)
-            } catch (ex: ConflictException) {
-                try {
-                    client.getSubject(authState, GetSubjectParser(authState)).let { authState ->
-                        updateSources(authState)
-                        sources[source.sourceId]?.let { source ->
-                            success(authState, source)
-                        } ?: failure(IllegalStateException("Source was not added to ManagementPortal, even though conflict was reported."))
-                    }
-                } catch (ioex: IOException) {
-                    logger.error("Failed to register source {} with {}: already registered",
-                            source.sourceName, source.type, ex)
-
-                    failure(ex)
-                }
-            } catch (ex: java.lang.IllegalArgumentException) {
-                logger.error("Source {} is not valid", source)
-                failure(ex)
-            } catch (ex: AuthenticationException) {
-                listener.invalidate(authState.token, false)
-                logger.error("Authentication error; failed to register source {} of type {}",
-                        source.sourceName, source.type, ex)
-                failure(ex)
-            } catch (ex: IOException) {
-                logger.error("Failed to register source {} with {}",
-                        source.sourceName, source.type, ex)
-                failure(ex)
-            } catch (ex: JSONException) {
-                logger.error("Failed to register source {} with {}",
-                        source.sourceName, source.type, ex)
-                failure(ex)
-            }
-        } ?: failure(IllegalStateException("Cannot register source without a client"))
-
-        return true
-    }
-
-
-    private fun updateSources(authState: AppAuthState) {
-        authState.sourceMetadata
-                .forEach { sourceMetadata ->
-                    sourceMetadata.sourceId?.let {
-                        sources[it] = sourceMetadata
-                    }
-                }
-    }
 
     override fun onDestroy() {
         config.config.removeObserver(configUpdateObserver)
@@ -266,10 +147,10 @@ class ManagementPortalLoginManager(private val listener: AuthService, state: App
                 config.getString(OAUTH2_CLIENT_ID),
                 config.getString(OAUTH2_CLIENT_SECRET, ""),
             )
-        } catch (e: MalformedURLException) {
+        } catch (_: MalformedURLException) {
             logger.error("Cannot construct ManagementPortalClient with malformed URL")
             null
-        } catch (e: IllegalArgumentException) {
+        } catch (_: IllegalArgumentException) {
             logger.error("Cannot construct ManagementPortalClient without client credentials")
             null
         }
@@ -286,42 +167,18 @@ class ManagementPortalLoginManager(private val listener: AuthService, state: App
         }
     }
 
-    private fun addSource(authState: AppAuthState, source: SourceMetadata): AppAuthState {
-        val sourceId = source.sourceId
-        return if (sourceId == null) {
-            logger.error("Cannot add source {} without ID", source)
-            authState
-        } else {
-            sources[sourceId] = source
-
-            authState.alter {
-                val existing = sourceMetadata.filterTo(HashSet()) { it.sourceId == source.sourceId }
-                if (existing.isEmpty()) {
-                    invalidate()
-                } else {
-                    sourceMetadata -= existing
-                }
-                sourceMetadata += source
-            }
-        }
-    }
-
-    private fun removeSource(authState: AppAuthState, source: SourceMetadata): AppAuthState {
-        sources.remove(source.sourceId)
-        return authState.alter {
-            val existing = sourceMetadata.filterTo(HashSet()) { it.sourceId == source.sourceId }
-            if (existing.isNotEmpty()) {
-                sourceMetadata -= existing
-                invalidate()
-            }
-            source.sourceId = null
-        }
-    }
-
     companion object {
         const val SOURCE_TYPE = "org.radarcns.auth.portal.ManagementPortal"
         private val logger = LoggerFactory.getLogger(ManagementPortalLoginManager::class.java)
         val sourceTypeList = listOf(SOURCE_TYPE)
+
+        @OptIn(ExperimentalContracts::class)
+        private fun ensureMpClient(client: AbstractRadarPortalClient): Boolean {
+            contract {
+                returns(true) implies (client is ManagementPortalClient)
+            }
+            return client is ManagementPortalClient
+        }
     }
 
     private data class ManagementPortalConfig(
